修正

+++ app/controllers/products_controller.rb
@@ -1,6 +1,6 @@
 class ProductsController < ApplicationController
   before_action :authenticate_user!, only: [:new, :edit]
-  before_action :set_product, only: [:show, :edit]
+  before_action :set_product, only: [:show, :edit, :destroy]
   before_action :move_to_index, only: [:edit]
 
 
@@ -42,8 +42,10 @@ class ProductsController < ApplicationController
 
   def destroy
     product = Product.find(params[:id])
-    product.destroy
-    redirect_to root_path
+    if product.user_id == current_user.id
+      product.destroy
+      redirect_to root_path
+    end
   end
 
   private
